+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+* Bottom tab bar now is hidden when:
+  - Typing (so is not above the keyboard)
+  - When navigating inside the stack (e.g. subject form)
 
 
 ## 1.4.0 - 2019-05-20